closure.py reports 0 symbols, 0 bytes for the class.  applyFrameAction is
BYTE-IDENTICAL at 287; process matches at 362 with free differences only;
applyAction (388 vs 387) and advanceTrellis (1038 vs 1054) differ by the same
spill the blob makes and we do not, examined and declined under 614's free
column rather than chased.

THE RESULT WORTH KEEPING is a type, not a symbol.  `getMetric` returned a
rounded `float` where the object returns its x87 accumulator unnarrowed, and
that changed which candidate the trellis commits -- the candidates come in
exactly-negating pairs with bit-identical metrics, so the tie is decided by
whether the stored `float` rounded up.  Return type is now `long double`.  The
file's own comment had predicted this and named advanceTrellis as the trigger.
Six spellings were compiled through the period compiler: the two that emit the
blob's `fcoms` both insert a narrowing and FAIL the differential test.  Findings
5854 and 5855, with the open question named rather than closed.

actionLookupTable's rule -- each entry's decimal digits are 1 + 2*previous +
current -- regenerates all 128 entries, and the test checks the regeneration
against the blob's own .data rather than against ours.

D800 RENUMBERED TO D930.  The branch measured the deviation high-water when it
was cut; v92-modulator-tail took D800 for V92Modulator::resamplerPhaseOffset
while it ran.  Two live entries, both correct when measured.  Renumbered only
the branch's -- its three references are V90SpectralShaper.cpp x2 and
t_v90spectrellis.cpp, while master's belong to V92Modulator.* and
t_v92modstate.cpp -- and the entry records what it used to be called.

tools/gccdiverge.json resolved by three-way set logic against the merge base,
not by taking a side: the branch's copy is 88 commits old and still declared
t_v90leaves and t_v90adid, which master SPLIT to make their mutation suites
scoreable.  Taking it would have reverted 617 recovered verdicts.  Eight entries
now; the branch's own t_v90specproc carries forward.  Note the branch's report
said "tools/ untouched", which was not accurate -- gccdiverge is a gate register
and it added an entry to it.  The entry is sound: V90SpectralVerifier::process
inherits t_psd's excess precision through Psd::process (1453), 181 of 35,450
GCC-13 checks and all of them spectrum bins, make period green, no mutation
suite registered.

+++ include/dsplib/V90SpectralShapingFilter.h
@@ -73,8 +73,38 @@ public:
 	 * Returns the accumulator, and is `const`: it runs the same recurrence
 	 * over `blocks * blockLength` samples in x87 registers and writes not
 	 * one word of the object back.
+	 *
+	 * `long double` AND NOT `float`, WHICH IS A BEHAVIOURAL FACT AND NOT A
+	 * PREFERENCE.  A C++ mangling carries no return type, so nothing in
+	 * `_ZNK24V90SpectralShapingFilter9getMetricEPKsj` decides this; the
+	 * object does.  It ends `fstp %st(1)` three times and returns
+	 * (0x33270) with NO NARROWING ON THE RETURN PATH, so the value reaching
+	 * the caller has the x87 stack's full 64-bit significand.  A `float`
+	 * return narrows it through memory -- measured, with the explicit cast
+	 * and without it, and both emit `fstps`/`flds` -- which the object does
+	 * not do here.
+	 *
+	 * IT DOES SPILL INTERNALLY, and that is not the same thing.  There are
+	 * two `fstps` inside the loop (0x3320d, 0x33213) with a matching
+	 * `flds` (0x33230), so some value IS going through a four-byte slot.
+	 * It cannot be the accumulator or the running state: declaring all of
+	 * those `float` makes this function's own differential suite fail 180
+	 * checks of 866, so the recurrence is carried at extended precision and
+	 * the spilled values are ones a float slot holds exactly -- the
+	 * coefficients, which are `float` members to begin with.
+	 *
+	 * IT IS ONLY OBSERVABLE TO A COMPARISON, AND `V90SpectralShaper::
+	 * advanceTrellis` MAKES ONE.  That function keeps its running best as
+	 * a `float` in memory and compares the freshly returned value against
+	 * it with `fcoms` BEFORE rounding it, so the two operands are at
+	 * different precisions on purpose.  Rounding here collapses that and
+	 * changes which trellis candidate wins an exact tie -- which is not a
+	 * hypothetical: with a zeroed filter state the search's candidates come
+	 * in exactly-negating pairs whose metrics are bit-identical, and the
+	 * tie is broken by whether the stored `float` rounded up or down.
+	 * Finding 5854.
 	 */
-	float	getMetric(const short *in, unsigned int blocks) const;
+	long double getMetric(const short *in, unsigned int blocks) const;
 
 	/* Public for offsetof; see V90ConstellationDesigner.h. */
 	float		coeff[4];	/* +0x00 setFilterCoeff's arguments  */
